requirements.txt
flask
Jinja2
arango
python-dotenv
python-arango
requests
psycopg2-binary
pandas
bs4 
BeautifulSoup4